chore: performances API 요청 관련 파일 하나로 통일
