--- src/apis/domains/performance/api.ts
@@ -1,94 +1,0 @@
-import { get, post, put } from "@apis/index";
-import { components } from "@typings/api/schema";
-import { ApiResponseType } from "@typings/commonType";
-import axios, { AxiosResponse } from "axios";
-
-export type PerformanceDetailResponse = components["schemas"]["PerformanceDetailResponse"];
-
-export const getPerformanceDetail = async (
-  performanceId: number
-): Promise<PerformanceDetailResponse | null> => {
-  try {
-    const response: AxiosResponse<ApiResponseType<PerformanceDetailResponse>> = await get(
-      `/performances/detail/${performanceId}`
-    );
-
-    return response.data.data;
-  } catch (error) {
-    console.error("error", error);
-    return null;
-  }
-};
-
-type BookingPerformanceDetailResponse = components["schemas"]["BookingPerformanceDetailResponse"];
-
-export const getBookingPerformanceDetail = async (
-  performanceId: number
-): Promise<BookingPerformanceDetailResponse | null> => {
-  try {
-    const response: AxiosResponse<ApiResponseType<BookingPerformanceDetailResponse>> = await get(
-      `/performances/booking/${performanceId}`
-    );
-
-    return response.data.data;
-  } catch (error) {
-    console.error("error", error);
-    return null;
-  }
-};
-
-type TicketAvailabilityResponse = components["schemas"]["TicketAvailabilityResponse"];
-
-export const getScheduleAvailable = async (
-  scheduleId: number,
-  purchaseTicketCount: number
-): Promise<TicketAvailabilityResponse | number> => {
-  try {
-    const response: AxiosResponse<ApiResponseType<TicketAvailabilityResponse>> = await get(
-      `/schedules/${scheduleId}/availability?purchaseTicketCount=${purchaseTicketCount}`
-    );
-
-    return response.data.data;
-  } catch (error) {
-    console.error("error", error);
-
-    if (axios.isAxiosError(error)) {
-      console.error("err is", error);
-      const errorStatus = error.response?.data.status;
-      console.log(errorStatus);
-
-      return errorStatus;
-    }
-    return -1;
-  }
-};
-
-export type PerformanceResponse = components["schemas"]["PerformanceResponse"];
-
-export const postPerformance = async (formData): Promise<PerformanceResponse | number> => {
-  try {
-    const response = await post("/performances", formData);
-
-    return response.data;
-  } catch (error) {
-    console.error("error", error);
-
-    return null;
-  }
-};
-
-export type PerformanceUpdateResponse = components["schemas"]["PerformanceUpdateResponse"];
-
-export const updatePerformance = async (
-  formData
-): Promise<PerformanceUpdateResponse | null | any> => {
-  try {
-    const response = await put("/performances", formData);
-
-    return response.data;
-  } catch (error) {
-    console.error("error", error);
-
-    return null;
-  }
-};
